requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
yt-dlp>=2023.7.6
pydub>=0.25.1
pydantic>=2.8.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.21.0
soundfile>=0.12.0
